requirements.txt
Flask==3.1.2
flask-cors==6.0.2
requests==2.32.5
beautifulsoup4==4.14.3
gunicorn==23.0.0
google-generativeai